+++ avm/ptn/alz/empty/CHANGELOG.md
@@ -20,6 +20,7 @@ The latest version of the changelog can be found [here](https://github.com/Azure
 ### Changes
 
 - Initial version
+- Updated ReadMe with AzAdvertizer reference
 
 ### Breaking Changes