+++ avm/res/app/job/CHANGELOG.md
@@ -19,6 +19,7 @@ The latest version of the changelog can be found [here](https://github.com/Azure
 - Updated resource provider
 - Added new `identity` property to the JobScaleRule type
 - Updated RP versions in tests
+- Updated ReadMe with AzAdvertizer reference
 
 ### Breaking Changes