+++ avm/res/db-for-my-sql/flexible-server/CHANGELOG.md
@@ -20,6 +20,7 @@ The latest version of the changelog can be found [here](https://github.com/Azure
 - New `configurations` parameter allows setting MySQL server configuration parameters directly through the module
 - Created new configuration child resource module at `configuration/main.bicep`
 - Added `configurationType` definition with proper typing for configuration parameters
+- Updated ReadMe with AzAdvertizer reference
 
 ### Breaking Changes